+++ fixture/react-native/react-native.config.js
@@ -1,0 +1,12 @@
+const path = require("path");
+
+const root = path.resolve(__dirname, "../../");
+
+module.exports = {
+  dependencies: {
+    "@shopify/flash-list": {
+      root,
+      platforms: {},
+    },
+  },
+};